functions: correct get_os logic
at commit aea49c3: functions: Fix logic for get_os() function
To determine which OS the scripts are running on, check if the
build.prop file exists instead of using the lsb_release command.
If it does, then the scripts are running on Android; otherwise
it's Ubuntu.

From the changelog, we can know get_os() expect to return 2(means android) if
build.prop file exists, but the code does not do that previouslly

Signed-off-by: Li Zhijian <lizhijian@cn.fujitsu.com>